+++ kernel/headers/procesos.h
@@ -9,7 +9,7 @@
 
 /////////////////////////////// Prototipos ///////////////////////////////
 void mostrar_pcb(t_pcb PCB); // Muestra por pantalla los valores de un PCB
-void mostrar_metrica(const char* nombre, uint16_t* metrica); // Auxiliar para mostrar_pcb
+void mostrar_metrica(const char* nombre, int* metrica); // Auxiliar para mostrar_pcb
 void mostrar_colas_estados(void);
 const char* estado_to_string(Estados estado); // Asocia el valor del enum con el nombre del estado (Para imprimir el nombre en vez de un numero)
 void cambiar_estado_pcb(t_pcb* PCB, Estados nuevo_estado_enum);
